Moving shared code to Common project in preparation for TriggerService access to CosmosDB.
