+++ src/Common/TesModels/TesTaskLog.cs
@@ -18,7 +18,7 @@ using System.Runtime.Serialization;
 using System.Text;
 using Newtonsoft.Json;
 
-namespace TesApi.Models
+namespace Common.TesModels
 {
     /// <summary>
     /// TaskLog describes logging information related to a Task.
