+++ src/Common/TesModels/TesTaskLogExtended.cs
@@ -8,7 +8,7 @@ using System.Reflection;
 using System.Runtime.Serialization;
 using Newtonsoft.Json;
 
-namespace TesApi.Models
+namespace Common.TesModels
 {
     /// <summary>
     /// TaskLog describes logging information related to a Task.
