+++ src/TesApi.Tests/CachingWithRetriesAzureProxyTests.cs
@@ -4,13 +4,13 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Common.TesModels;
 using LazyCache;
 using LazyCache.Providers;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using Polly.Utilities;
-using TesApi.Models;
 using TesApi.Web;
 
 namespace TesApi.Tests
