+++ src/TesApi.Tests/DeleteCompletedBatchJobsHostedServiceTests.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Linq;
 using System.Threading.Tasks;
+using Common;
+using Common.TesModels;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
-using TesApi.Models;
 using TesApi.Web;
 
 namespace TesApi.Tests
