+++ src/TesApi.Web/ContractResolvers/MinimalTesTaskContractResolver.cs
@@ -4,9 +4,9 @@
 using System;
 using System.Collections.Generic;
 using System.Reflection;
+using Common.TesModels;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
-using TesApi.Models;
 
 namespace TesApi.Web
 {
