+++ src/TesApi.Web/DeleteCompletedBatchJobsHostedService.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Common;
+using Common.TesModels;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using TesApi.Models;
 
 namespace TesApi.Web
 {
