+++ src/TesApi.Web/IAzureProxy.cs
@@ -5,8 +5,8 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Common.TesModels;
 using Microsoft.Azure.Batch;
-using TesApi.Models;
 
 namespace TesApi.Web
 {
