+++ src/TesApi.Web/TesApi.Web.csproj
@@ -39,4 +39,7 @@
       <_Parameter1>TesApi.Tests</_Parameter1>
     </AssemblyAttribute>
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Common\Common.csproj" />
+  </ItemGroup>
 </Project>
